requirements.txt
click==8.1.7
cmake==3.28.3
dlib==19.24.99
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.2.0
PyMySQL==1.1.0
python-dotenv==1.0.1
